requirements.txt
flask
gunicorn
google-generativeai
python-dotenv
requests
flask-compress